+++ DistributedTransactions.Domain/Orders/RebalancingRequest.cs
@@ -2,7 +2,7 @@ namespace DistributedTransactions.Domain.Orders;
 
 public class RebalancingRequest : OrdersDomainModel
 {
-    public required int RequestNumber { get; set; }
-    public required int GroupNumber { get; set; }
+    public required long RequestNumber { get; set; }
+    public required long GroupNumber { get; set; }
     public DateTime Date { get; set; }
 }
